backend/.ipynb_checkpoints/requirements-checkpoint.txt
flask
flask-cors
openai
volcengine-python-sdk
faiss-cpu
gdown
sentence-transformers
numpy
gunicorn
huggingface_hub